Version 2.12.0

+++ Compressor.m
@@ -187,7 +187,7 @@
       default:
         DebugLog(0, @"Error while decompressing data : Insufficient memory" );
         DebugLog(0, @"destLen = %u", destLen);
-        DebugLog(0, @"destLenf = %u", destLenf);
+        DebugLog(0, @"destLenf = %u", (unsigned int)destLenf);
         DebugLog(0, @"error = %d", error);
         break;
     }//end switch(error)
